+++ lib/generators/server_shelf.dart
@@ -2,25 +2,23 @@
 // All rights reserved. Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-import '../stagehand.dart';
 import '../src/common.dart';
 import 'server_shelf_data.dart';
 
-/**
- * A generator for a server app built on `package:shelf`.
- */
+/// A generator for a server app built on `package:shelf`.
 class ServerShelfGenerator extends DefaultGenerator {
   ServerShelfGenerator()
       : super('server-shelf', 'Web Server',
             'A web server built using the shelf package.',
             categories: const ['dart', 'shelf', 'server']) {
-    for (TemplateFile file in decodeConcatenatedData(data)) {
+    for (var file in decodeConcatenatedData(data)) {
       addTemplateFile(file);
     }
 
     setEntrypoint(getFile('bin/server.dart'));
   }
 
-  String getInstallInstructions() => "${super.getInstallInstructions()}\n"
-      "run your app via 'dart ${entrypoint.path}'";
+  @override
+  String getInstallInstructions() => '${super.getInstallInstructions()}\n'
+      'run your app via `dart ${entrypoint.path}`.';
 }