+++ lib/generators/web_simple.dart
@@ -2,19 +2,16 @@
 // All rights reserved. Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-import '../stagehand.dart';
 import '../src/common.dart';
 import 'web_simple_data.dart';
 
-/**
- * A generator for a uber-simple web application.
- */
+/// A generator for a uber-simple web application.
 class WebSimpleGenerator extends DefaultGenerator {
   WebSimpleGenerator()
       : super('web-simple', 'Bare-bones Web App',
             'A web app that uses only core Dart libraries.',
             categories: const ['dart', 'web']) {
-    for (TemplateFile file in decodeConcatenatedData(data)) {
+    for (var file in decodeConcatenatedData(data)) {
       addTemplateFile(file);
     }