+++ lib/generators/web_stagexl.dart
@@ -2,19 +2,16 @@
 // All rights reserved. Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-import '../stagehand.dart';
 import '../src/common.dart';
 import 'web_stagexl_data.dart';
 
-/**
- * A generator for a StageXL web application.
- */
+/// A generator for a StageXL web application.
 class WebStageXlGenerator extends DefaultGenerator {
   WebStageXlGenerator()
       : super('web-stagexl', 'StageXL Web App',
             'A starting point for 2D animation and games.',
             categories: const ['dart', 'web']) {
-    for (TemplateFile file in decodeConcatenatedData(data)) {
+    for (var file in decodeConcatenatedData(data)) {
       addTemplateFile(file);
     }