+++ templates/web-angular/lib/src/todo_list/todo_list_component.dart
@@ -1,9 +1,6 @@
-// Copyright (c) __year__, __author__. All rights reserved. Use of this source code
-// is governed by a BSD-style license that can be found in the LICENSE file.
-
 import 'dart:async';
 
-import 'package:angular2/angular2.dart';
+import 'package:angular/angular.dart';
 import 'package:angular_components/angular_components.dart';
 
 import 'todo_list_service.dart';
@@ -31,7 +28,11 @@ class TodoListComponent implements OnInit {
     items = await todoListService.getTodoList();
   }
 
-  void add(String description) => items.add(description);
+  void add() {
+    items.add(newTodo);
+    newTodo = '';
+  }
+
   String remove(int index) => items.removeAt(index);
   void onReorder(ReorderEvent e) =>
       items.insert(e.destIndex, items.removeAt(e.sourceIndex));